fix: not accounting for collapsed rules
AWS likes to collapse several IpPermission objects into a single
one with multiple IpRange entries when the FromPort, ToPort and
IpProtocol fields are identical. As a quick naive solution that
works, the IpPermission rules are expanded such that each IpRange
object becomes its own IpPermission object and inherits the parent
fields. This restores the old assumption that each IpRule is
its own IpPermission.